+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to Stride Pro will be documented in this file.
 
 ## [1.0.1] - 2026-05-02
 
+### New Features
+
+- Added user avatar upload functionality - users can now upload custom profile pictures to replace the default avatar image
+
 ### Bug Fixes
 
 - Fixed issue with controlling music playback